gc3libs.auth.grid: Make `arc_flavour` a module-level global.
This looks better than exposing it in the signature of module public
functions, since the value is determined at module import time and
should not be changed.

On the other hand, since some functionality is still missing/not
working in ARC1, there's a need to override the value in order to fall
back on the ARC0 method.  So the override is still listed in a public
function's signature, but I hope its role is more clear now.

  modified:   grid.py