Make cPickle work with all the `Store` classes preserving the feature of saving different `Persistable` objects to different files or db records.
In cPickle `Pickler` and `Unpickler` are actually functions which return a class, therefore it is impossible to inherit from Pickle in order to override the functions persistent_id and persistent_load, needed in order to store different Persistable to different records.

In order to achieve the same result, with cPickle we need to create two functions: `persistent_load` and `persistent_id`, which will be then assigned to the class returned by `cPickle.Pickler()` or `cPickle.Unpickler()`, but we also need to access a reference to the `Store` class in these functions.

The proposed solution, thus, is to create two callable classes which are initialized with the current `Store` instance, so that the previous behavior is preserved.