MemoryPool class added.
MemoryPool is a class used to store Proxies and call the "proxy_forget" method of thoose proxies we don't want/cannot keep in memory.

To decide which proxies have to be forgetted you can either pass to the constructor a function which will decide which object have to stay in memory and which one have to be saved, or you can specify the maximum number of objects that can be kept in memory.