`SqlStore`: extra_fields does not accept strings as keys anymore.
Cleanup of SqlStore constructor. Now we accept only a valid sqlalchemy.Column instance as key of the extra_fields dictionary. In addiction, we always create the schema of the table starting from *our* vision of what should it be, without trying to merge the current schema as it appear from sqlalchemy.MetaData().reflect() with what the caller expects to have.

This will basically move any error coming from an inconsistency between the current schema and the application to other part of the code, but since this is basically a *configuration* error we must assume that the schema passed is correct.

Moreover, a database with more fields than expected should not be an issue, and I don't see why we should care about fields the caller don't care either.

	modified:   gc3libs/persistence/sql.py
	modified:   gc3libs/persistence/tests/test_persistence.py