Merge branch 'sql'.
Too many changes to list (again) here, but the bottom line is:
a session now has a configurable storage location, which can be
a directory on the filesystem (FilesystemStore, the default so far)
or can be a table in an SQL database (of any kind supported by SQLAlchemy).

Ah, and a session is now a "real" object in Python code, not some
"figure of speech" / "ad-hoc code in SessionBasedScript".