First working implementation of a quantity system.
This allows the memory and time requests to be specified using a
unit-aware system. On the programmer side, quantities can be specified
with a relatively simple syntax:

  >>> requested_walltime = 8*hours # or`days`, `minutes`, `seconds`
  >>> requested_memory = 2*GB # or `TB`, `MB`, `kB`

On the user side, parser and conversion-to-string functions allow for
human-compatible input and output of those values, e.g.,

  >>> requested_walltime = Duration('1 hour 30 minutes')

See all the relevant documentation in `gc3libs/quantity.py`.