Application: Raise an `InvalidValue` error if input/output file names cannot be converted to character strings.
Fixes Issue 312.

This should be revisited, but indeed there is a problem if we get a
UNICODE string because -by definition- it cannot be converted without
specifying an encoding, which we do know nothing about.  The solution,
as discussed in Issue 159, should be not to have UNICODE at all in
GC3Pie, and this is a step in that direction.