Agree on a few common attributes set at job end by each backend.
See Issue 78 for details.

Update Issue 78
After a job reaches the `TERMINATED` state, the following
attributes are also set.  The meaning and format of the attributes
is consistent across backends.

  `execution.duration`
    Time lapse from start to end of the job at the remote
    execution site, as a `gc3libs.quantity.Duration`:class: value.
    (This is also often referred to as the 'wall-clock time' or
    `walltime`:term: of the job.)

  `execution.max_used_memory`
    Maximum amount of RAM used during job execution, represented
    as a `gc3libs.quantity.Memory`:class: value.

  `execution.used_cpu_time`
    Total time (as a `gc3libs.quantity.Duration`:class: value) that the
    processors has been actively executing the job's code.

Backends may set other attributes as well; the only convention is that
the name of the attribute starts with the (lowercased) backend name.
For instance, the PbsLrms backend sets attributes `pbs_queue`,
`pbs_end_time`, etc.

The triple (submission time, actual start time, end time) would be a
very useful addition to this set of common attributes, but not all
backends provide this information. (PBS, SGE and ARC1 do; ARC0 does
not have the "actual start time").  What would be a good way to handle
this?