ARC0: Perform exclusion of already-used resubmission targets at "execution target" level, not "queue" level.
Fixes Issue 346.

The problem in Issue 346 originated from the fact that the exclusion
happened too early: there can be many queues, but only one of them
matches the requirements.  Thus, excluding that one queue led to an
empty list of targets later on.  If on the other hand we do exclusion
on the basis of targets, then we can detect if the target list is
empty because of our exclusion method and reset it.