[DRAFT] Update EC2Lrms to find the "best" IP address to connect to a VM.
When updating the subresource associated to a VM, both IP address (public_dns_name and private_ip_address) are tried. If the current "preferred_ip" (by default, private_ip_address) is not working (i.e. we get a TransportError when updating the subresource), then the other is tried. The "correct" one is then saved in the VMPool instance file, and loaded as "preferred_ip" when VMPool is later initialized.

This change is aimed to allow gc3pie to run inside a VM in Hobbes, where it's currently impossible to access the VM using its public IP address, but it's possible to connect to it using the private one.

The change needs revision, as in its current state is sub-optimal and possibly buggy!

	modified:   gc3libs/backends/ec2.py
	modified:   gc3utils/commands.py