Make `bacct` the default `acct` command for `LsfLrms`.
Also check issue 384.

Some times ago we moved the default acct command from `bacct` to
`bjobs` because of some issues with some LSF clusters.

However, the current implementation was not able to correctly parse
the output of `bjobs`, and collect the correct values for `duration`,
`used_cpu_time` and `max_used_memory`.