Remove dependency of OpenStack and EC2 backends on the `gc3libs.config`.
The code in those backends used to instanciate a
`gc3libs.config.Configuration()` object with the sole purpose of using
it to create `ShellcmdLrms` instances using whatever was passed in the
`vm_auth` config key as the `auth` key of the `ShellcmdLrms`.  This
approach had a big problem: one needs to re-read exactly the same set
of configuration files for the vm_auth<->auth correspondence to work.
However, the location of the configuration files used was not
available in the `OpenStackLrms` or `Ec2Lrms`; hence, they would just
hard-code the std configuration path location, thus opening the door
to a hard-to-trace bug.

Instead, we now have the `Configuration` object treat `vm_auth`
specially, just like `auth`, compile it into a function object, and
make `OpenStackLrms` and `Ec2Lrms` take an additional *mandatory*
argument `vm_auth`.  Whatever is passed as that argument is then
re-used to create `ShellcmdLrms` subresources.  And since subresources
are always `ShellcmdLrms`, there's no need to invoke the
`Configuration._make_resource()` factory: we can just call the
`ShellcmdLrms` constructor directly.