Rewrite `update_job_status()` in batch-system backends.
This is a large rewrite, prompted by a corner case in SLURM's handling
of the job exit status (see Issue #507) but the logic in that function
had become too hairy to be amenable to a simple fix.

Hence this rewrite for cleanup purposes:

* standardize the return type of the `_parse_stat_output()` to
  a named tuple, with documentation for all the fields.

* compute the "termination status" in the `_parse_acct_output()`
  functions, instead of passing back exit code and (sometimes) a signal
  number, with varying field names.

* ensure that all `_parse_acct_output()` functions return a dictionary
  with at standardized field names (`termstatus` + common resource usage
  names from Issue #78 + backend-specific fields).

I'm also quite sure that the "accounting delay" functionality had been
broken in the older code, and this commit should fix that.