New implementation of daemon service using XML-RPC, with client
This implementation has a big advantage: it relies on SimpleXMLRPCServer and
xmlrpclib, which are included in the Python Standard Library. The disadvantage
is that instead of using unix socket we need to use a TCP socket as the standard
implementation only allow XML-RPC over TCP.

The port is chosen automatically and then written to a file in the working directory.

Each SessionDaemonScript now accepts a --client option to interact with the corresponding daemon.