Create the comm daemon outside its thread
An hypotetical user could want to add some method to the communication
thread. However, this was not always possible due to a race condition:
if the code in `before_main_loop` in the SessionBasedDaemon children
class is executed before the thread actually created the _CommDaemon
class the script will fail.