Fix an issue when killing and resubmitting a SequentialTaskCollection
When a task collection is killed and then redo() is called, the
follwing error happens:

Traceback (most recent call last):
  File "/home/antonio/.virtualenvs/gc3pie/local/lib/python2.7/site-packages/nose/case.py", line 197, in runTest
    self.test(*self.arg)
  File "/home/antonio/git/gc3pie/gc3libs/tests/test_bug_kill_redo.py", line 49, in test_kill_redo
    engine.redo(seq, from_stage=0)
  File "/home/antonio/git/gc3pie/gc3libs/core.py", line 1679, in redo
    task.redo(*args, **kwargs)
  File "/home/antonio/git/gc3pie/gc3libs/workflow.py", line 394, in redo
    super(SequentialTaskCollection, self).redo()
  File "/home/antonio/git/gc3pie/gc3libs/__init__.py", line 569, in redo
    .format(self, self.execution.state))
AssertionError: Can only re-do a Task which is in a terminal state; task <helpers.SimpleSequentialTaskCollection object at 0x7f4eef455690> is in state RUNNING instead.

The root cause is that the kill() method will not change the state of
all the inner tasks to TERMINATED, so the task will still have a
RUNNING state, and the redo() refuses to kill jobs in non-terminal
state.