Add comment on kernel parameters that should be checked when using INotifyPoller
When using inotify there are certain kernel parameters that might be
increased increased in order to allow a reasonably large workflow to
work. These are:

* fs.inotify.max_user_instances
* fs.inotify.max_user_watches
* fs.inotify.max_queued_events

Defaults are pretty low, and if they are not increased the application
will receive a "too many open files" error which is not so easy to
troubleshoot.