Update `*_instance_id` attributes to `_lrms_vm_id` on load.
This fixes commit b43ad26 so that existing sessions and tasks are not
broken by the commit but are transparently converted (using the hook
introduced in commit a1e7893).

The new name is consistent with the naming convention adopted so far,
that object attributes that are only needed for backends' internal use
begin with a `_lrms_` prefix.