Drop support for the OpenStack backend on Python 2.6
Upstream has ceased supporting Python 2.6 since `novaclient` version 3.0.0; at
the time of this writing, we're past 7.0 already and it's really become
impossible to keep code compatible across so many major revisions.  Since we're
dropping support for Python 2.6 in GC3Pie soon-ish, start by ceasing support for
the OpenStack backend on Python 2.6.

Installation of the optional feature `[openstack]` will *not* error out (which
could be regarded as a bug, but I could not figure out how to detect when it's
being installed from within `setup.py` and return a meaningful error message),
but any attempt to use an `OpenStackLrms` resource will result in a
`ImportError` (which automatically disables the resource).