Always look up a "front end" name in the SSH config file and possibly translate it.
If a "HostName" directive is specified in the SSH config file, the host name or
IP address following it are used in stead of the "frontend =" value from the
GC3Pie config file.

This was supposed to happen since SSH config lookups were implemented, but a
wrong code path resulted in overwriting the output of the "HostName" directive
with the value read in GC3Pie's config file.

Thanks to Kyle Robertson for reporting the bug!