Make `Session.load()` also add a task to the session by default.
The behavior is overridable with the new `add=...` argument.

In addition, make `Session.forget()` be silently successfull when
passed the ID of a task that is *not* in the session.  When modifying
existing code to use the new API and distinguish cases when we want a
task added or not, this seems necessary to have some removal cases
handled successfully.