Fix counting of `ok` and `failed` pseudo-states.
The issue with counting, as implemented through the pub/sub pattern,
is that the final exit code may be decided by any of the handlers that
are subscribed to the task state changes -- and calling order of
subscribed handlers is not guaranteed.

The fix requires that we allow an `Engine` instance to subscribe to
exitcode/returncode changes as well, and then modify the counts if the
task final returncode is changed.