Allow loading tasks persisted with pre-2.6 releases of GC3Pie.
When trying to load old task files or data persisted with pre-2.6
releases (before the compatibility with Python 3), an error like this
will happen:

```
  File ".../gc3libs/persistence/sql.py", line 397, in load
    obj = make_unpickler(self, BytesIO(rawdata[0])).load()
  File "/usr/lib/python2.7/pickle.py", line 864, in load
    dispatch[key](self)
  File "/usr/lib/python2.7/pickle.py", line 1230, in load_build
    d = inst.__dict__
AttributeError: 'UrlValueDict' object has no attribute '__dict__'
```

The fix involves overwriting a the `Unpickler.load_build()` method
with a patched version that uses `setattr()` when `__dict__` is not
available (as is the case in GC3Pie's `UrlKeyDict` and `UrlValueDict`
since they are now derived from `namedtuple`).  This may be brittle
and break again if the actual implementation of Python's `pickle` changes.

The fix is only implemented in Python2, since Py3 has a completely
different implementation of `pickle`.  What mitigates this limitation
is that old session data in unlikely to occur in Python3 code (which
was not supported in pre-2.6).