Correctly handle conversion of empty bytes objects
The original intention of this code is to convert bytes objects into
strings.

However, the `if stdout` and `if stderr` tests originally employed here
fail when stdout or stderr is a zero length/empty bytes object i.e.
when equal to `b''`. These tests fail because an empty bytes object
evaluates to False. So empty bytes objects are currently not converted
to strings.

The failure only occurs in Python 3. In Python 3, bytes is a different
type than string whereas in Python 2 the bytes class is simply an
alias for the string class.