Allow passing an SSH private key to`SshAuth` (#665)
Allow specification of a private key when constructing an `SshAuth` object.

If specified, it will be used for authentification to the remote resource, 
much like the `pkey` argument to `paramiko.client.SSHClient.connect()`. 
This is useful for any case where the key is already in memory,  as 
writing it to a file can be avoided.