Cirrus CI: do not run session-based script tests.
They seem to stall.  My guess is some Linux functionality is not available in the
unprivileged containers used by Cirrus CI, so it should go away by switching
the unit tests to run in a VM instead of a Linux container, but for the time
being it's quicker to just disable the tests.