Fix doctests for `gc3libs.utils.check_file_access`.
Doctest assumed non-root access, which is not true for example when running
the test suite in a Docker container.

They still assume the existence of a POSIX file system (`/bin/sh`, `/tmp`) and
possibly Linux-only files (`/proc/version`) but I cannot imagine how to write
a doctest in a more system-independent way.