Fix docstring example in `SshTransport`'s constructor.
Constructing an Engine from a programmatically-given configuration requires
that also a *resource* section is present; we must give it in the supplied
configuration dict, otherwise the test depends on the presence of a valid
configuration file in the standard locations (which might not be the case when
running e.g. in a Docker container).