Cirrus CI: Ensure Python is installed before running "Install Script Tests".
This is a bit of a shot in the dark: as far as I understand, Cirrus CI does
not provide a way of specifying the order in which scripts should be executed,
and just runs them in the order they are defined in the YAML file.  This makes
it a bit complicated to use templates, as the order of definition of `_script`
items depends on the order templates are instantiated.  I hope that putting
`<<: &TEMPLATE` after a script inserts the inherited scripts after the script
just defined...