[DATALAD RUNCMD] Fix some of the typos with codespell
This is far from a complete set of fixes. I just spotted a typo in README.md and decided
to share opportunity of using codespell to help with fixing and then automate CI to alert when
typos are introduced.

Have look e.g. at
https://github.com/heartexlabs/label-studio/pull/3515
for an example on how to add codespell workflow to github CI etc.

=== Do not change lines below ===
{
 "chain": [],
 "cmd": "bash -c 'codespell -L keypair,keypairs -w ./gc3libs/optimizer/TODO README.rst ./gc3libs/etc/gspg_wrapper.py ./gc3libs/tests/test_config.py *py ./docs/contributors/*.rst ./gc3libs/backends/*.py || :'",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^